# docs/swagger.yaml
basePath: /api/v1
definitions:
  response.Data:
    properties:
      code:
        $ref: '#/definitions/response.ResCode'
      data: {}
      msg: {}
    type: object
...
  schema.UserListResponse:
    properties:
      list:
        description: User list | 用户列表
        items:
          $ref: '#/definitions/schema.UserListItem'
        type: array
      page:
        description: Current page number | 当前页码
        type: integer
      page_size:
        description: Items per page | 每页数量
        type: integer
      total:
        description: Total count | 总数量
        type: integer
    type: object
...
  schema.UserUpdateRequest:
    properties:
      avatar:
        description: Avatar URL | 头像URL
        example: https://example.com/avatar.jpg
        type: string
      email:
        description: Email | 邮箱
        example: test@example.com
        type: string
      id:
        description: User ID | 用户ID
        example: 1
        type: integer
      readme:
        description: README content | README内容
        example: |-
          ## 关于我
          这是我的README内容
        type: string
      signature:
        description: Signature | 个性签名
        example: 这是我的个性签名
        type: string
      username:
        description: Username | 用户名
        example: testuser
        maxLength: 100
        minLength: 3
        type: string
    required:
    - id
    type: object
...
host: localhost:9876
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: PokeForum is a forum application based on Gin framework | PokeForum
    是一个基于 Gin 框架的论坛应用程序
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PokeForum API
  version: "1.0"
paths:
...
  /manage/users:
    get:
      consumes:
      - application/json
      description: Get paginated user list with keyword search and status filtering
        support | 分页获取用户列表，支持关键词搜索和状态筛选
      parameters:
      - description: Page number | 页码
        in: query
        name: page
        required: true
        type: integer
      - description: Items per page | 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      - description: Search keyword | 搜索关键词
        example: '"test"'
        in: query
        name: keyword
        type: string
      - description: User status | 用户状态
        example: '"Normal"'
        in: query
        name: status
        type: string
      - description: User role | 用户身份
        example: '"User"'
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieve successful | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserListResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get user list | 获取用户列表
      tags:
      - '[Admin]User Management | [管理员]用户管理'
    post:
      consumes:
      - application/json
      description: Administrator creates new user account | 管理员创建新用户账户
      parameters:
      - description: User information | 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Creation successful | 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserDetailResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Create user | 创建用户
      tags:
      - '[Admin]User Management | [管理员]用户管理'
    put:
      consumes:
      - application/json
      description: Update user's basic information | 更新用户的基本信息
      parameters:
      - description: User information | 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update successful | 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserDetailResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Update user information | 更新用户信息
      tags:
      - '[Admin]User Management | [管理员]用户管理'
  /manage/users/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information for specified user | 获取指定用户的详细信息
      parameters:
      - description: User ID | 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Retrieve successful | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserDetailResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get user details | 获取用户详情
      tags:
      - '[Admin]User Management | [管理员]用户管理'
  /manage/users/balance/logs:
    get:
      consumes:
      - application/json
      description: Get paginated user balance change logs with various filtering options
        | 分页获取用户余额变动记录，支持多种筛选条件
      parameters:
      - description: Page number | 页码
        in: query
        name: page
        required: true
        type: integer
      - description: Items per page | 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      - description: User ID filter | 用户ID筛选
        in: query
        name: user_id
        type: integer
      - description: Change type filter | 变动类型筛选
        example: '"points"'
        in: query
        name: type
        type: string
      - description: Start date | 开始日期
        example: '"2024-01-01"'
        in: query
        name: start_date
        type: string
      - description: End date | 结束日期
        example: '"2024-12-31"'
        in: query
        name: end_date
        type: string
      - description: Operator ID filter | 操作者ID筛选
        in: query
        name: operator_id
        type: integer
      - description: Related business type filter | 关联业务类型筛选
        example: '"post"'
        in: query
        name: related_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieve successful | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserBalanceLogResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get user balance change logs | 获取用户余额变动记录
      tags:
      - '[Admin]User Management | [管理员]用户管理'
  /manage/users/balance/summary/{id}:
    get:
      consumes:
      - application/json
      description: Get balance summary statistics for specified user | 获取指定用户的余额汇总统计信息
      parameters:
      - description: User ID | 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Retrieve successful | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserBalanceSummary'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get user balance summary information | 获取用户余额汇总信息
      tags:
      - '[Admin]User Management | [管理员]用户管理'
  /manage/users/ban:
    post:
      consumes:
      - application/json
      description: Ban specified user, supports temporary and permanent bans | 封禁指定用户，支持短期封禁和永久封禁
      parameters:
      - description: Ban information | 封禁信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserBanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ban successful | 封禁成功
          schema:
            $ref: '#/definitions/response.Data'
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Ban user | 封禁用户
      tags:
      - '[Admin]User Management | [管理员]用户管理'
  /manage/users/currency:
    put:
      consumes:
      - application/json
      description: Add or deduct currency for user | 为用户增加或减少货币
      parameters:
      - description: Currency information | 货币信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserCurrencyUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update successful | 更新成功
          schema:
            $ref: '#/definitions/response.Data'
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Update user currency | 更新用户货币
      tags:
      - '[Admin]User Management | [管理员]用户管理'
  /manage/users/moderator/categories:
    put:
      consumes:
      - application/json
      description: Set the list of categories managed by specified moderator | 为指定版主设置其管理的版块列表
      parameters:
      - description: Category information | 版块信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.ModeratorCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Setting successful | 设置成功
          schema:
            $ref: '#/definitions/response.Data'
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Set moderator categories | 设置版主管理版块
      tags:
      - '[Admin]User Management | [管理员]用户管理'
  /manage/users/points:
    put:
      consumes:
      - application/json
      description: Add or deduct points for user | 为用户增加或减少积分
      parameters:
      - description: Points information | 积分信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserPointsUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update successful | 更新成功
          schema:
            $ref: '#/definitions/response.Data'
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Update user points | 更新用户积分
      tags:
      - '[Admin]User Management | [管理员]用户管理'
  /manage/users/role:
    put:
      consumes:
      - application/json
      description: Update user's role permissions (Regular User, Moderator, Admin,
        etc.) | 更新用户的身份权限（普通用户、版主、管理员等）
      parameters:
      - description: Role information | 身份信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserRoleUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update successful | 更新成功
          schema:
            $ref: '#/definitions/response.Data'
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Update user role | 更新用户身份
      tags:
      - '[Admin]User Management | [管理员]用户管理'
  /manage/users/status:
    put:
      consumes:
      - application/json
      description: Update user's status (Normal, Muted, Banned, etc.) | 更新用户的状态（正常、禁言、封禁等）
      parameters:
      - description: Status information | 状态信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserStatusUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update successful | 更新成功
          schema:
            $ref: '#/definitions/response.Data'
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Update user status | 更新用户状态
      tags:
      - '[Admin]User Management | [管理员]用户管理'
  /manage/users/unban:
    post:
      consumes:
      - application/json
      description: Remove ban status from specified user | 解除指定用户的封禁状态
      parameters:
      - description: Unban information | 解封信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserUnbanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Unban successful | 解封成功
          schema:
            $ref: '#/definitions/response.Data'
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Unban user | 解封用户
      tags:
      - '[Admin]User Management | [管理员]用户管理'
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"